pyproject.toml
[tool.poetry]
name = "xsltmapper"
version = "0.1.0"
description = ""
authors = ["Sotades <anthonyjbateman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
openpyxl = "^3.1.5"
langchain-openai = "^0.2.14"
langchain = "^0.3.13"
pytest = "^8.3.4"
langchain-community = "^0.3.13"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
